moved TeamHighlight and VariantPrompt
