+++ src/scenes/ProgressReports/Detail/ProgressReportDrawer/Steps/PromptVariant/index.tsx
@@ -1,0 +1,1 @@
+export * from './PromptVariantStep';
