+++ src/scenes/ProgressReports/Detail/ProgressReportDrawer/Steps/TeamHighlight/TeamHighlightStep.graphql
